Split StatsHandlerTest into its own file
